+++ package.json
@@ -16,6 +16,7 @@
     "@types/node": "^13.7.0",
     "@types/react": "^16.9.19",
     "@types/react-dom": "^16.9.5",
+    "@types/react-router-dom": "^5.1.2",
     "@types/styled-components": "^4.4.2",
     "@types/styled-system": "^5.1.6",
     "@types/styled-system__css": "^5.0.4",
