+++ src/App.test.tsx
@@ -5,5 +5,5 @@ import App from './App';
 
 test('renders learn react link', async () => {
   const { getByText } = render(<App />);
-  await waitForElement(() => getByText(/Riddle me this/i));
+  await waitForElement(() => getByText(/Home/i));
 });
