+++ src/App.tsx
@@ -1,54 +1,49 @@
-import React, { useState } from 'react';
-import { useEvents } from 'modules/events';
+import React from 'react';
+import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
+import Home from 'pages/home';
+import AddEvents from 'pages/add-events';
+import Auth from 'pages/auth';
+import Vote from 'pages/vote';
 
 const App = () => {
-  const { events, addEvent, removeEvent } = useEvents();
-  const [title, setTitle] = useState('');
-  const [description, setDescription] = useState('');
-  const [start, setStart] = useState('');
-  const [end, setEnd] = useState('');
-  const images: string[] = [];
-  const attendees: string[] = [];
-  const createEvent = () => {
-    addEvent({
-      title,
-      description,
-      start: new Date(start),
-      end: new Date(end),
-      images,
-      attendees,
-    });
-  };
   return (
-    <div style={{ margin: 20 }}>
-      <h1>List of events</h1>
-      <ul>
-        {events.map(event => (
-          <li key={event.id}>
-            <h2>{event.title}</h2>
-            <p>{event.description}</p>
-            <button onClick={() => removeEvent(event.id)}>Delete</button>
-          </li>
-        ))}
-      </ul>
-      <h1>Add event</h1>
+    <Router>
       <div>
-        <h3>Title</h3>
-        <input value={title} onChange={e => setTitle(e.target.value)}></input>
-        <h3>Description</h3>
-        <input
-          value={description}
-          onChange={e => setDescription(e.target.value)}
-        ></input>
-        <h3>Start date</h3>
-        <input value={start} onChange={e => setStart(e.target.value)}></input>
-        <h3>End date</h3>
-        <input value={end} onChange={e => setEnd(e.target.value)}></input>
+        <nav>
+          <ul>
+            <li>
+              <Link to="/">Home</Link>
+            </li>
+            <li>
+              <Link to="/auth">Sign up / Sign in</Link>
+            </li>
+            <li>
+              <Link to="/add-events">Add Events</Link>
+            </li>
+            <li>
+              <Link to="/vote">Vote</Link>
+            </li>
+          </ul>
+        </nav>
+
+        {/* A <Switch> looks through its children <Route>s and
+            renders the first one that matches the current URL. */}
+        <Switch>
+          <Route path="/add-events">
+            <AddEvents />
+          </Route>
+          <Route path="/auth">
+            <Auth />
+          </Route>
+          <Route path="/vote">
+            <Vote />
+          </Route>
+          <Route path="/">
+            <Home />
+          </Route>
+        </Switch>
       </div>
-      <div style={{ marginTop: 10 }}>
-        <button onClick={createEvent}>Create Event</button>
-      </div>
-    </div>
+    </Router>
   );
 };
