+++ src/pages/auth/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Auth = () => {
+  return <div></div>;
+};
+
+export default Auth;
