+++ src/pages/vote/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Vote = () => {
+  return <div></div>;
+};
+
+export default Vote;
